AWS/CDK/CustomPolicyForUseCodeCommit/requirements.txt
aws-cdk.core==1.204.0
aws-cdk.aws-iam==1.204.0